Editor menu reorganization
