+++ planet-clicker/Assets/LibplanetUnity/Editor/BlockChainMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 084ff49f7c4de354fa78822effd71d61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
