+++ planet-clicker/Assets/LibplanetUnity/Editor/GenesisBlockMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 16e7ef3dfd5727544903120b45c8328a
+guid: bb4f4cd12ba39dc40801a81134b4a4d0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
