+++ planet-clicker/Assets/LibplanetUnity/Editor/PrivateKeyMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 872f83ce86772c14999c94b66e76666e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
