+++ planet-clicker/Assets/LibplanetUnity/Editor/SwarmConfigMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec91880a6c8e5f04fba4df82fefed2c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
