+++ planet-clicker/Assets/LibplanetUnity/Editor/UtilsMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 49067befed5e10d4f8e73ef57fca53d8
+guid: a8a9dfb70ad1c5d42a707ab50ac8ab9f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
